Add guard clause to run activerecord error only when AR exist
