+++ lib/pry-byetypo/exceptions/active_record/uninitialized_constant.rb
@@ -1,25 +1,21 @@
 # frozen_string_literal: true
 
-require_relative "../exceptions_base"
+require_relative "base"
 
 module Exceptions
-  module NameError
-    class UninitializedConstant < ExceptionsBase
+  module ActiveRecord
+    class UninitializedConstant < Base
       private
 
       def unknown_from_exception
         exception.to_s.split.last
       end
 
-      def corrected_word
-        @corrected_word ||= spell_checker(ar_models_dictionary).correct(unknown_from_exception).first
-      end
-
       def corrected_cmd
         @corrected_cmd ||= last_cmd.gsub(unknown_from_exception, corrected_word)
       end
 
-      def ar_models_dictionary
+      def dictionary
         @ar_models_dictionary ||= store.transaction { |s| s["active_record_models"] }
       end
     end
