+++ lib/pry-byetypo/exceptions/exceptions_base.rb
@@ -29,7 +29,11 @@ class ExceptionsBase < Base
     @pry = pry
   end
 
-  def spell_checker(dictionary)
+  def corrected_word
+    @corrected_word ||= spell_checker.correct(unknown_from_exception).first
+  end
+
+  def spell_checker
     DidYouMean::SpellChecker.new(dictionary: dictionary)
   end
