+++ lib/pry-byetypo/setup/store.rb
@@ -1,6 +1,7 @@
 module Setup
   module Store
     DEFAULT_STORE_PATH = "byetypo_dictionary.pstore"
+    SEVEN_DAYS = 604800
 
     def store
       @store ||= PStore.new(store_path)
@@ -8,6 +9,14 @@ module Setup
 
     def store_path
       ENV["BYETYPO_STORE_PATH"] || DEFAULT_STORE_PATH
+    end
+
+    # By default we update the store every week.
+    # TODO: Make it configurable
+    def staled_store?
+      return true if store["synced_at"].nil?
+
+      (store["synced_at"] + SEVEN_DAYS) <= Time.now
     end
   end
 end
